SPN-558 - XAResource implementation in ISPN to become fully stateless
